+++ Readme.md
@@ -13,4 +13,5 @@
     - if bcrypt does not work or gives error then try using [bcrypt.js](https://www.npmjs.com/package/bcryptjs) or 
        bycrpt node.js
 - jsonwebToken (Generate webtoken )  
-- [Jwt.io](https://jwt.io/) (to extract the token information)    
+- [Jwt.io](https://jwt.io/) (to extract the token information)   
+- [Http Status Codes](https://restfulapi.net/http-status-codes/)
