+++ api/routes/orders.js
@@ -1,129 +1,19 @@
 const express = require('express');
 const router = express.Router();
-const mongoose = require('mongoose');
+const checkAuth = require("../middleware/check-auth")
 
-const Order = require('../models/order');
-const Product = require('../models/product');
-const checkAuth  = require("../middleware/check-auth")
+const OrdersController = require('../controllers/orders');
 
+//get all orders secured via api-key
+router.get('/', checkAuth, OrdersController.order_get_all)
 
-router.get('/',checkAuth , (req, res, next) => {
-    Order.find()
-    .select('product quantity _id')
-    .populate('product' ,'name')//instead of ide all object got populated
-    .exec()
-        .then(docs => {
-            res.status(200).json({
-                count: docs.length,
-                orders: docs.map((doc => {
-                    return {
-                        _id: doc.id,
-                        product: doc.product,
-                        quantity: doc.quantity,
-                        request: {
-                            type: 'GET',
-                            url: 'http://localhost:3000/orders/' + doc.id
-                        }
-                    }
+//post orders secured via api-key
+router.post("/", checkAuth, OrdersController.order_post_product);
 
-                })),
+//get single orders by id secured via api-key
+router.get('/:orderId', checkAuth, OrdersController.order_get_product);
 
-            })
-        })
-        .catch(err => {
-            res.status(500).json({
-                error: err
-            })
-        })
-
-});
-
-
-
-router.post("/",checkAuth , (req, res, next) => {
-    Product.findById(req.body.product)
-        .then(product => {
-            if (!product) {
-                return res.status(404).json({
-                    message: "Product not found"
-                });
-            }
-            const order = new Order({
-                _id: mongoose.Types.ObjectId(),
-                quantity: req.body.quantity,
-                product: req.body.product
-            });
-            return order.save();
-        })
-        .then(result => {
-            console.log(result);
-            res.status(201).json({
-                message: "Order stored",
-                createdOrder: {
-                    _id: result._id,
-                    product: result.product,
-                    quantity: result.quantity
-                },
-                request: {
-                    type: "GET",
-                    url: "http://localhost:3000/orders/" + result._id
-                }
-            });
-        })
-        .catch(err => {
-            console.log(err);
-            res.status(500).json({
-                error: err
-            });
-        });
-});
-
-router.get('/:orderId',checkAuth , (req, res, next) => {
-    Order.findById(req.params.orderId)
-    .populate('product')
-    .exec()
-    .then(order=>{
-        if(!order)
-        {
-            return res.status(400).json(
-                {
-                    message:'Order Not Found',
-                }
-            )
-        }
-        res.status(200).json({
-           order:order ,
-           request:{
-            type: "GET",
-            url: "http://localhost:3000/orders"
-        }
-        })
-    })
-    .catch((err=>{
-        res.status(500).json({
-            error:err
-        })
-    }))
-    
-});
-
-router.delete('/:orderId',checkAuth , (req, res, next) => {
-    Order.remove({
-        _id:req.params.orderId
-    })
-    .exec()
-    .then(order=>{
-        res.status(200).json({
-           message:'Order Deleted',
-           request:{
-            type: "POST",
-            url: "http://localhost:3000/orders",
-            body:{product:'ID',quantity:'Number'}
-        }
-        })
-    })
-    .catch();
-    
-});
+//delete single orders by id secured via api-key
+router.delete('/:orderId', checkAuth, OrdersController.order_delete_product);
 
 module.exports = router;
