+++ api/routes/products.js
@@ -1,11 +1,11 @@
 const express = require('express');
 const router = express.Router(); //register Route
 
-const Product = require('../models/product');
-
-const mongoose = require('mongoose');
 const multer = require('multer');
 const checkauth = require('../middleware/check-auth')
+const ProductController = require('../controllers/products');
+
+//setting up the storage folder
 const storage = multer.diskStorage({
     destination: function(req,file,cb)
     {
@@ -15,180 +15,36 @@ const storage = multer.diskStorage({
     cb(null,new Date().getUTCMilliseconds() + file.originalname)
     }
 });
+
+//accepting the specified like png and jpeg
 const filefilter =(req,file,cb)=>{
     if(file.mimetype === 'image/jpeg'||file.mimetype ==='image/png')
     {cb(null,true);}
     else{
         cb(null,false);
-    }
-    
+    } 
 }
+
+//size should be less than 5 mb
 const upload = multer({storage:storage,limits:{
     fileSize : 1024 * 1024 * 5
 },
 fileFilter:filefilter});//folder path for multer to store images
 
- 
-router.get('/', (req, res, next) => {
-    Product.find()
-        .select('name price _id productImage') //select fields you want to get
-        .exec()
-        .then(docs => {
-            if (docs) {
-                const response = {
-                    count: docs.length,
-                    products: docs.map(
-                        doc => {
-                            return {
-                                name: doc.name,
-                                price: doc.price,
-                                _id: doc._id,
-                                productImage:doc.productImage,
-                                request: {
-                                    type: 'GET',
-                                    url: 'http://localhost:3000/products/' + doc.id
-                                }
-                            }
-                        }
-                    )
-                }
-                res.status(200).json({
-                    response
-                })
+ //get all products 
+router.get('/', ProductController.get_all_products);
 
-            } else {
-                res.status(404).json({
-                    message: "No data Found"
-                })
-            }
-        });
-});
+//post single product
+router.post("/",checkauth,upload.single('productImage'), ProductController.post_product);
 
+//get product by id
+router.get('/:productId', ProductController.get_product_by_id);
 
-//upload.single() is a multer object for a single file only
-router.post("/",checkauth,upload.single('productImage'), (req, res, next) => {
-   console.log(req.file)
-    const product = new Product({
-        _id: new mongoose.Types.ObjectId(),
-        name: req.body.name,
-        price: req.body.price,
-        productImage:req.file.path
-    });
-    product
-        .save()
-        .then(result => {
-            console.log(result);
-            res.status(201).json({
-                message: "Created Product succesfully",
-                createdProduct: {
-                    name: result.name,
-                    price: result.price,
-                    _id: result._id,
-                    request: {
-                        type: 'POST',
-                        url: 'http://localhost:3000/products/' + result.id
-                    }
-                }
-            })
-        })
-        .catch(err => {
-            console.log(err);
-            res.status(500).json({
-                error: err
-            });
-        });
-});
+//update product with id and param value
+router.patch('/:productId',checkauth, ProductController.update_product_by_id);
 
-//adding parametre to get
-router.get('/:productId', (req, res, next) => {
-    const id = req.params.productId;
-
-    Product.findById(id)
-        .select('name price _id productImage')
-        .exec()
-        .then(doc => {
-            console.log(doc)
-            if (doc) {
-                res.status(200).json({
-                    product: doc,
-                    request: {
-                        type: 'GET',
-                        url: 'http://localhost:3000/products'
-                    }
-                })
-            } else {
-                res.status(404).json({
-                    message: "No valid Data found or id not found"
-                })
-            }
-        }).
-    catch(err => {
-        console.log(err)
-        res.status(500).json({
-            error: err
-        })
-    });
-});
-
-router.patch('/:productId',checkauth, (req, res, next) => {
-    const id = req.params.productId;
-    const updateOps = {};
-    for (const ops of req.body) {
-        updateOps[ops.propName] = ops.value;
-    }
-    Product.update({
-        _id: id
-    }, {
-        $set: updateOps
-    }).exec().then(result => {
-        console.log('\n' + result + 'update')
-        res.status(200).json({
-            message: 'Product Updated',
-            data: result,
-            request: {
-                type: 'GET',
-                url: 'http://localhost:3000/products/' + id
-            }
-        })
-    }).catch(
-        err => {
-            console.log(err)
-            res.status(500).json({
-                message: err
-            })
-        }
-    );
-});
-
-router.delete('/:productId',checkauth, (req, res, next) => {
-    const id = req.params.productId;
-    Product.remove({
-            _id: id
-        })
-        .exec()
-        .then(result => {
-            console.log('Hello' + result)
-            res.status(200).json({
-                message: 'Product deleted',
-                requests: {
-                    type: 'POST',
-                    url: 'http://localhost:3000/products/',
-                    body: {
-                        name: 'String',
-                        price: 'Number'
-                    }
-                }
-                // deleted:result
-            })
-        })
-        .catch(err => {
-            console.log(err)
-            req.status(500).json({
-                error: err
-            })
-
-        })
-});
+//delete product by id
+router.delete('/:productId',checkauth,ProductController.delete_product_by_id);
 
 //necessary for using route
 module.exports = router;
